Initial commit of reduced rank jacobian project
Initial commit with status to date. This includes the following
directories and their contents:
     base_run    includes data from the no-perturbation run
     input       includes all generated input, including the
                 estimated and true Jacobians
     paper       includes an old draft that needs to be
                 almost entirely rewritten
     python      includes python files that create the input
                 files, as well as inversion.py, which includes
                 three inherited classes: Inversion,
                 ReducedRankInversion, which inherits from
                 Inversion, and ReducedRankJacobian, which
                 inherits from ReducedRankInversion.
     run_dir     includes the bash scripts that actually run the
                 python scripts

Notably, this commit represents progress made up until the inclusion
of the Broyden method to estimating the Jacobian. The inversion.py
file includes the Broyden method applied to the perturbation of
eigenvectors, which has been debugged, and the perturbation of an
aggregated grid, which has not been debugged. There are also commented
out sections that reflect previous attempts at recovering the
Jacobian. These commented lines will be deleted in future commits since
they reflect attempts that did not work.